+++ tests/Unit/PHPUnitTest.php
@@ -3,6 +3,7 @@
 namespace Realodix\NextProject\Test;
 
 use PHPUnit\Framework\AssertionFailedError;
+use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
 use Realodix\NextProject\Test\Fixtures\ObjectEquals\ValueObject;
 
@@ -11,6 +12,8 @@ use Realodix\NextProject\Test\Fixtures\ObjectEquals\ValueObject;
 
 final class PHPUnitTest extends TestCase
 {
+    use PHPUnitTestProvider;
+
     public function testStringContainsString(): void
     {
         ass('foo bar')
@@ -565,6 +568,51 @@ final class PHPUnitTest extends TestCase
         ass('<foo/>')
             ->xmlStringToString('<foo/>')
             ->xmlStringNotToString('<bar/>');
+    }
+
+    /**
+     * @dataProvider stringContainsStringIgnoringLineEndingsProvider
+     *
+     * @param string $needle
+     * @param string $haystack
+     */
+    public function testStringContainsStringIgnoringLineEndings(string $needle, string $haystack): void
+    {
+        ass($haystack)
+            ->stringContainsStringIgnoringLineEndings($needle);
+    }
+
+    public function testNotStringContainsStringIgnoringLineEndings(): void
+    {
+        $this->expectException(ExpectationFailedException::class);
+
+        ass("\r\nc\r\n")
+            ->stringContainsStringIgnoringLineEndings("b\nc");
+    }
+
+    /**
+     * @dataProvider stringEqualIgnoringLineEndingsProvider
+     *
+     * @param string $expected
+     * @param string $actual
+     */
+    public function testStringEqualIgnoringLineEndings(string $expected, string $actual): void
+    {
+        ass($actual)
+            ->stringEqualIgnoringLineEndings($expected);
+    }
+
+    /**
+     * @dataProvider stringEqualIgnoringLineEndingsFailProvider
+     *
+     * @param string $expected
+     * @param string $actual
+     */
+    public function testNotStringEqualIgnoringLineEndings(string $expected, string $actual): void
+    {
+        $this->expectException(ExpectationFailedException::class);
+        ass($actual)
+            ->stringEqualIgnoringLineEndings($expected);
     }
 }
